Schematic object encapsulation and other minor improvements.
* Encapsulate file name member of base screen object.
* Encapsulate associated screen member of schematic sheet object.
* Create add screen method to schematic sheet object to simplify setting
  the associated screen.
* Move the change file name code in the schematic sheet object to the edit
  sheet method in the schematic editor frame object to eliminate message
  dialogs.
* Improve reference counting in schematic screen object.
* Add helper type definitions for changing schematic object storage to C++
  containers.

+++ pcbnew/printout_controler.cpp
@@ -264,7 +264,7 @@ void BOARD_PRINTOUT_CONTROLER::DrawPage()
          * is the y coordinate values from  - PlotAreaSize.y to 0 */
 #ifdef USE_WX_ZOOM
         int y_dc_offset = PlotAreaSizeInPixels.y;
-        y_dc_offset = (int) y_dc_offset * userscale;
+        y_dc_offset = (int) ( ( double ) y_dc_offset * userscale );
         dc->SetDeviceOrigin( 0, y_dc_offset );
 #endif
         int ysize = (int) ( PlotAreaSizeInPixels.y / scaley );
